finalizando recebaaaaaaaaaaaa

+++ Py/extrair_texto.py
@@ -2,27 +2,33 @@ import cv2
 import pytesseract
 import schedule
 import time
+import os
 
 #C:\Program Files\Tesseract-OCR
-
+cont = 0
 def rodar():
 
-    imagem = cv2.imread('receba.jpg')
+    if(os.path.exists(r'C:\Users\rafag\OneDrive\Documentos\PDS_projeto\System-Cont\Py\receba.jpg')):
 
-    pytesseract.pytesseract.tesseract_cmd = "C:\\Program Files\\Tesseract-OCR\\tesseract.exe"
+        global cont
+        cont = cont+1
 
-    texto = pytesseract.image_to_string(imagem)
+        imagem = cv2.imread('receba.jpg')
 
-    with open("texto_extraido.doc", "w") as arquivo_doc:
-        arquivo_doc.write(texto)
+        pytesseract.pytesseract.tesseract_cmd = "C:\\Program Files\\Tesseract-OCR\\tesseract.exe"
 
-    import os
+        texto = pytesseract.image_to_string(imagem)
 
-    list_arquivos = os.listdir(r"C:\Users\rafag\OneDrive\Documentos\PDS_projeto\System-Cont\Py")
+        with open("texto_extraido.doc", "w") as arquivo_doc:
+            arquivo_doc.write(texto)
 
-    for arquivo in list_arquivos:
-        if "texto_extraido.doc" in arquivo:
-            os.rename(f"C:\\Users\\rafag\\OneDrive\\Documentos\\PDS_projeto\\System-Cont\\Py\\{arquivo}", f"C:\\Users\\rafag\\OneDrive\\Documentos\\PDS_projeto\\System-Cont\\Files\\Finished\\{arquivo}")
+
+        list_arquivos = os.listdir(r"C:\Users\rafag\OneDrive\Documentos\PDS_projeto\System-Cont\Py")
+        for arquivo in list_arquivos:
+            if "texto_extraido.doc" in arquivo:
+                os.rename(f"C:\\Users\\rafag\\OneDrive\\Documentos\\PDS_projeto\\System-Cont\\Py\\texto_extraido.doc", f"C:\\Users\\rafag\\OneDrive\\Documentos\\PDS_projeto\\System-Cont\\Files\\Finished\\texto_extraido{cont}.doc")
+        os.remove('receba.jpg')
+        print(texto)
 
 schedule.every(5).seconds.do(rodar)
 
@@ -31,4 +37,3 @@ while True:
     print("runing")
     time.sleep(1)
 
-print(texto)
